BVH working but not extremely efficient

+++ src/main/scala/bounding/ShapeContainer.scala
@@ -6,6 +6,6 @@ import renderer.Hit
 trait ShapeContainer {
   def size : Int
   def intersect(ray: Ray): Option[Hit]
-  def intersectionTest(r: Ray, maxDist: Float): Boolean
+  def intersectionTest(r: Ray, maxDist: Double): Boolean
 
 }
