+++ src/main/scala/material/Material.scala
@@ -43,13 +43,13 @@ object Material {
 }
 
 final case class UnshadedColor(color: RGB,
-                               ambient: Float,
-                               diffuse: Float,
-                               spec: Float,
-                               reflective: Float,
-                               refractive: Float,
-                               n: Float,
-                               shininess: Float)
+                               ambient: Double,
+                               diffuse: Double,
+                               spec: Double,
+                               reflective: Double,
+                               refractive: Double,
+                               n: Double,
+                               shininess: Double)
 
 trait Material {
   def getMat(position: Vector3): UnshadedColor
